Route all saved export task rebuilds through one place
