+++ corehq/apps/export/utils.py
@@ -6,8 +6,6 @@ from datetime import datetime
 from dimagi.utils.couch.undo import DELETED_SUFFIX
 from dimagi.utils.modules import to_function
 from dimagi.utils.couch import CriticalSection
-from soil import DownloadBase
-from soil.progress import get_task_status
 from toggle.shortcuts import set_toggle
 
 from corehq.apps.accounting.utils import domain_has_privilege
@@ -691,30 +689,3 @@ def domain_has_excel_dashboard_access(domain):
 
 def domain_has_daily_saved_export_access(domain):
     return domain_has_privilege(domain, DAILY_SAVED_EXPORT)
-
-
-def _get_saved_export_download_data(export_instance_id):
-    download_id = 'rebuild_export_tracker.{}'.format(export_instance_id)
-    download_data = DownloadBase.get(download_id)
-    if download_data is None:
-        download_data = DownloadBase(download_id=download_id)
-    return download_data
-
-
-def saved_export_set_task(celery_task_result, export_instance_id):
-    """Associates a celery task with a particular export instance
-
-    :param celery_task_result: the result of my_celery_task.delay(..)
-    """
-    download_data = _get_saved_export_download_data(export_instance_id)
-    download_data.set_task(celery_task_result)
-
-
-def get_saved_export_task_status(export_instance_id):
-    """Get info on the ongoing rebuild task if one exists.
-
-    (This is built with the assumption that there shouldn't be multiple
-    rebuilds in progress for a single export instance)
-    """
-    download_data = _get_saved_export_download_data(export_instance_id)
-    return get_task_status(download_data.task)
